# docs/swagger.yaml
basePath: /api/v1
definitions:
  Backend_internal_models.CandidateResponse:
    properties:
      class:
        example: "2024"
        type: string
      created_at:
        example: "2025-11-12T03:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      major:
        example: informatics
        type: string
      mission:
        example: Serve all students equally
        type: string
      name:
        example: John Doe
        type: string
      profile_picture:
        example: https://example.com/image.jpg
        type: string
      updated_at:
        example: "2025-11-12T03:00:00Z"
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      vision:
        example: My vision for better education
        type: string
      vote_count:
        example: 5
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pufacomputing.ac.id
    name: API Support
    url: https://compsci.president.ac.id/support
  description: Backend API for PUFA Computer Science with Candidates and Votes Management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://compsci.president.ac.id/terms
  title: PUFA Backend API
  version: "1.0"
paths:
...
  /candidates:
    get:
      consumes:
      - application/json
      description: Get a paginated list of all candidates with optional filters by
        major and class
      parameters:
      - description: Filter by major (information_system or informatics)
        in: query
        name: major
        type: string
      - description: Filter by class
        in: query
        name: class
        type: string
      - default: 1
        description: Page number for pagination
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of candidates" example({"success":true,"data":[{"id":1,"name":"John
            Doe","vision":"My vision","mission":"My mission","class":"2024","major":"informatics","profile_picture":"https://...","vote_count":5}],"totalResults":1,"totalPages":1})
          schema:
            additionalProperties: true
            type: object
      summary: List all candidates
      tags:
      - Candidates
  /candidates/{candidateID}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific candidate including vote
        count
      parameters:
      - description: Candidate ID
        in: path
        name: candidateID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Candidate details
          schema:
            properties:
              data:
                $ref: '#/definitions/Backend_internal_models.CandidateResponse'
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Invalid candidate ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Candidate not found
          schema:
            additionalProperties: true
            type: object
      summary: Get candidate by ID
      tags:
      - Candidates
  /candidates/{candidateID}/delete:
    delete:
      consumes:
      - application/json
      description: 'Delete a candidate by ID. Requires Admin role (role_id: 1) and
        candidate:delete permission.'
      parameters:
      - description: Bearer token (Admin only)
        in: header
        name: Authorization
        required: true
        type: string
      - description: Candidate ID
        in: path
        name: candidateID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Candidate deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid candidate ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin role required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Candidate not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Delete a candidate (Admin only - role_id: 1)'
      tags:
      - Candidates
  /candidates/{candidateID}/edit:
    put:
      consumes:
      - multipart/form-data
      description: 'Update candidate information with optional profile picture upload.
        Requires Admin role (role_id: 1) and candidate:edit permission.'
      parameters:
      - description: Bearer token (Admin only)
        in: header
        name: Authorization
        required: true
        type: string
      - description: Candidate ID
        in: path
        name: candidateID
        required: true
        type: integer
      - description: Candidate data as JSON string
        example: '{"name":"John Doe","vision":"Updated vision","major":"informatics"}'
        in: formData
        name: data
        required: true
        type: string
      - description: New profile picture
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Candidate updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin role required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Candidate not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Update a candidate (Admin only - role_id: 1)'
      tags:
      - Candidates
  /candidates/create:
    post:
      consumes:
      - multipart/form-data
      description: 'Create a new candidate with profile picture upload. Requires Admin
        role (role_id: 1) and candidate:create permission.'
      parameters:
      - description: Bearer token (Admin only)
        in: header
        name: Authorization
        required: true
        type: string
      - description: Candidate data as JSON string
        example: '{"name":"John Doe","vision":"My vision","mission":"My mission","class":"2024","major":"informatics"}'
        in: formData
        name: data
        required: true
        type: string
      - description: Profile picture
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Candidate created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Admin role required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 'Create a new candidate (Admin only - role_id: 1)'
      tags:
      - Candidates
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"